+++ quam/components/hardware.py
@@ -140,6 +140,20 @@ class FrequencyConverter(BaseFrequencyConverter):
     def LO_frequency(self):
         return self.local_oscillator.frequency
 
+    @LO_frequency.setter
+    def LO_frequency(self, value):
+        """Sets the frequency of the local oscillator object"""
+        if self.local_oscillator is None:
+            raise AttributeError(
+                f"Unable to set LO frequency for {self} as it has no local oscillator"
+            )
+
+        # Use set_at_reference to ensure the frequency is updated, even if the local
+        # oscillator frequency is a reference
+        self.local_oscillator.set_at_reference(
+            "frequency", value, allow_non_reference=True
+        )
+
     def configure(self):
         if self.local_oscillator is not None:
             self.local_oscillator.configure()